Naively turn off UseOfMfc
In addressing the complaint here,
https://www.hard-light.net/forums/index.php?topic=80935.msg1853522#msg1853522

I have decided to include all the standard libraries for the debug build in the same way that the retail build does. I don't have enough information to know why this is different, but I could not resolve my build issues by instead adding more required libraries to the "Linked Dependencies". I could retry this solution later and roll this back.