Add a Text packing utility to netcode
Collapses a lot of warnings down to a single utility method. At the same time reduces a bit of duplication.

I've found this impossible to test so far as I cannot seem to run two instances of Starshatter locally. This change should result in identical behaviour so is low-risk but should be revisited once multiplayer is turned back on.